% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDSeroSurv.R
\name{FMDSeroSurv}
\alias{FMDSeroSurv}
\title{This package is developed for the estimation of state & national level estimators and the related measures for sero-surveilance.}
\usage{
FMDSeroSurv(SeroSurvData, Census_Data)
}
\arguments{
\item{SeroSurvData}{NSP-Ab antibody n X 3 data frame (n: number of states in the sample) obtained from sero-surveillance, where row represents the states, first column represents the total sample collected from each state and third column represents the number of DIVA positive samples.}

\item{Census_Data}{Animal census N X 1 data frame (state wise bovine population), where rows are states and column is the bovine population (cattle + buffalo) (e.g. N = 28 before 2014 & N= 29 after 2014)}
}
\value{
A data frame containing the state-level and national level estimates of DIVA positivity rates and related parameters of FMDV infection:
\itemize{
  \item1 DIVA_Positive, Var_DIVA_Prop., StError_Prop., MarginErr_Prop., CI_Lower, CI_Upper, Pred_Total are the DIVA positive rates (proportion as estimator), variance of the estimator, standard error, margin of error, 95 percentage CI and predicted total of animals having history FMD infection at the individual state-level.
  \item2 DIVA_Pos_Prop, Pos Rate, StError, CV, MarError, Lower_CI, Upper_CI, PredNatTotal, SE_PredTotal are the DIVA positive proportion, rates, its standard error, margin of error, co-efficient of variation, 95 percentage CI, predicted total, standard error of predicted total at the country/national level.
 }

 @author Samarendra Das

 @example
 set.seed(23)
 SeroSurvData <- data.frame(round(runif(5, 100, 200)), rnbinom(5, mu = 40, size = 1), row.names = c("AP", "KN", "MH", "OD", "Guj"))
 colnames(SeroSurvData) <- c("Samples", "Positive")
 Census_Data <- data.frame(round(runif(10, 400, 600)), row.names=c("AP", "KN", "MH", "OD", "Guj", "Goa", "Raj", "Assam", "Miz", "Meg"))
 colnames(Census_Data) <- "Total"
 #Do not run
 result <- FMDSeroSurv (SeroSurvData, Census_Data)
 state_estimate <- result$StateEstimates
 Nat_estimate <- result$NationalEstimates
}
\description{
This package is developed for the estimation of state & national level estimators and the related measures for sero-surveilance.
}
